main: Move fix_macos_nswindow_tabbing to utils
