+++ orangecanvas/gui/utils.py
@@ -4,7 +4,11 @@ Helper utilities
 """
 import sys
 import traceback
+from typing import List
+
 import ctypes
+import ctypes.util
+import platform
 
 from contextlib import contextmanager
 from typing import Optional, Union
@@ -175,6 +179,59 @@ def is_dwm_compositing_enabled():  # type: () -> bool
     rval = DwmIsCompositionEnabled(ctypes.byref(enabled))
 
     return rval == 0 and enabled.value
+
+
+def macos_set_nswindow_tabbing(enable=False):
+    # type: (bool) -> None
+    """
+    Disable/enable automatic NSWindow tabbing on macOS Sierra and higher.
+
+    See QTBUG-61707
+    """
+    if sys.platform != "darwin":
+        return
+    ver, _, _ = platform.mac_ver()
+    ver = tuple(map(int, ver.split(".")[:2]))
+    if ver < (10, 12):
+        return
+
+    c_char_p, c_void_p = ctypes.c_char_p, ctypes.c_void_p
+    id = Sel = Class = c_void_p
+
+    def annotate(func, restype, argtypes):
+        func.restype = restype
+        func.argtypes = argtypes
+        return func
+    try:
+        libobjc = ctypes.cdll.LoadLibrary(ctypes.util.find_library("libobjc"))
+        # Load AppKit.framework which contains NSWindow class
+        # pylint: disable=unused-variable
+        AppKit = ctypes.cdll.LoadLibrary(ctypes.util.find_library("AppKit"))
+        objc_getClass = annotate(
+            libobjc.objc_getClass, Class, [c_char_p])
+        objc_msgSend = annotate(
+            libobjc.objc_msgSend, id, [id, Sel])
+        sel_registerName = annotate(
+            libobjc.sel_registerName, Sel, [c_char_p])
+        class_getClassMethod = annotate(
+            libobjc.class_getClassMethod, c_void_p, [Class, Sel])
+    except (OSError, AttributeError):
+        return
+
+    NSWindow = objc_getClass(b"NSWindow")
+    if NSWindow is None:
+        return
+    setAllowsAutomaticWindowTabbing = sel_registerName(
+        b'setAllowsAutomaticWindowTabbing:'
+    )
+    # class_respondsToSelector does not work (for class methods)
+    if class_getClassMethod(NSWindow, setAllowsAutomaticWindowTabbing):
+        # [NSWindow setAllowsAutomaticWindowTabbing: NO]
+        objc_msgSend(
+            NSWindow,
+            setAllowsAutomaticWindowTabbing,
+            ctypes.c_bool(enable),
+        )
 
 
 def gradient_darker(grad, factor):
