+++ orangecanvas/main.py
@@ -11,7 +11,6 @@ import logging
 import optparse
 import pickle
 import shlex
-import shutil
 import io
 from urllib.request import getproxies
 from contextlib import ExitStack, redirect_stdout, redirect_stderr, closing
@@ -29,67 +28,13 @@ from .application.outputview import TextStream, ExceptHook
 
 from . import utils, config
 from .gui.splashscreen import SplashScreen
+from .gui.utils import macos_set_nswindow_tabbing as fix_macos_nswindow_tabbing
+
 from .registry import qt
 from .registry import WidgetRegistry, set_global_registry
 from .registry import cache
 
 log = logging.getLogger(__name__)
-
-
-def fix_macos_nswindow_tabbing():
-    """
-    Disable automatic NSWindow tabbing on macOS Sierra and higher.
-
-    See QTBUG-61707
-    """
-    import ctypes
-    import ctypes.util
-    import platform
-
-    if sys.platform != "darwin":
-        return
-    ver, _, _ = platform.mac_ver()
-    ver = tuple(map(int, ver.split(".")[:2]))
-    if ver < (10, 12):
-        return
-
-    c_char_p, c_void_p = ctypes.c_char_p, ctypes.c_void_p
-    id = Sel = Class = c_void_p
-
-    def annotate(func, restype, argtypes):
-        func.restype = restype
-        func.argtypes = argtypes
-        return func
-    try:
-        libobjc = ctypes.cdll.LoadLibrary(ctypes.util.find_library("libobjc"))
-        # Load AppKit.framework which contains NSWindow class
-        # pylint: disable=unused-variable
-        AppKit = ctypes.cdll.LoadLibrary(ctypes.util.find_library("AppKit"))
-        objc_getClass = annotate(
-            libobjc.objc_getClass, Class, [c_char_p])
-        objc_msgSend = annotate(
-            libobjc.objc_msgSend, id, [id, Sel])
-        sel_registerName = annotate(
-            libobjc.sel_registerName, Sel, [c_char_p])
-        class_getClassMethod = annotate(
-            libobjc.class_getClassMethod, c_void_p, [Class, Sel])
-    except (OSError, AttributeError):
-        return
-
-    NSWindow = objc_getClass(b"NSWindow")
-    if NSWindow is None:
-        return
-    setAllowsAutomaticWindowTabbing = sel_registerName(
-        b'setAllowsAutomaticWindowTabbing:'
-    )
-    # class_respondsToSelector does not work (for class methods)
-    if class_getClassMethod(NSWindow, setAllowsAutomaticWindowTabbing):
-        # [NSWindow setAllowsAutomaticWindowTabbing: NO]
-        objc_msgSend(
-            NSWindow,
-            setAllowsAutomaticWindowTabbing,
-            ctypes.c_bool(False),
-        )
 
 
 def fix_win_pythonw_std_stream():
